More header adjustments and white space fix.
Continued work on the Header section of the page. Added more HTML to the header logo to turn it into a link. Adjusted the sizing of the logo so it is more in line with the actual homepage. added missing styles to the form on desktop devices. Fixed a bug that would generate whitespace on smaller devices. Which was caused by containers in the footer not having a width assigned.