Latest news section work.
Applied font sizes to title and link elements and applied icomoon arrows and gave spacing. Changed hover effect to only apply to the article containers and not items within. Removed white space between the images and article block containers. Applied more media queries to the article block containers so they all remain the same size in all viewports. Added extra padding to the bottom of the section in smaller viewports.