Footer element and header adjustments
Made the icons more vertically aligned with their button containers by setting a top value to the icons. Applied this method to the search icon in the header section of the page to make the icon more vertically aligned. Compared the footer section to live and adjusted the font size, font weight and line height for the UL and LI elements. Adjusted the padding, font size, font weight and line height for elements within the sticky footer to be more inline with the actual homepage.