Small fixes
Adjusted the font size and weight of the about section titles to be more inline with the actual homepage. Tested against live. Reviewed the spacing above and below the view all link at the bottom of the latest news section. Removed unneeded margin from the view all link to correct the spacing between elements and to be more in line with the actual home. Adjusted the vertical alignment of the top view all link to be in the centre of its container and to be aligned with the title.